Sửa đc lỗi

+++ InsBook/Common/Accessories.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Numerics;
 using System.Web;
 
 namespace InsBook.Common
@@ -32,13 +31,13 @@ namespace InsBook.Common
               .Select(s => s[random.Next(s.Length)]).ToArray());
         }
 
-        public BigInteger GetTime()
+        public UInt64 GetTime()
         {
             var utc0 = new DateTime(2020, 2, 20, 0, 0, 0, 0, DateTimeKind.Utc); // thời gian mốc
             var issueTime = DateTime.Now; // thời gian hiện tại
 
             // lấy thời gian hiện tại trừ thời gian mốc -> đổi ra giây
-            var iat = (BigInteger)issueTime.Subtract(utc0).TotalSeconds;
+            var iat = (UInt64)issueTime.Subtract(utc0).TotalSeconds;
             return iat;
         }
     }
